+++ lib/csscss/parser/css.rb
@@ -36,12 +36,30 @@ module Csscss
           space?
         }
 
+        rule(:mixin_attributes) {
+          (
+            str('/* CSSCSS START MIXIN') >>
+            (str('*/').absent? >> any).repeat >>
+            str('*/') >>
+            (str('/* CSSCSS END MIXIN').absent? >> any).repeat >>
+            str('/* CSSCSS END MIXIN') >>
+            (str('*/').absent? >> any).repeat >>
+            str('*/') >>
+            space?
+          ).as(:mixin)
+        }
+
         rule(:ruleset) {
           (
             match["^{}"].repeat(1).as(:selector) >>
             str("{") >>
             space? >>
-            (comment | attribute | blank_attribute).repeat(0).as(:properties) >>
+            (
+              mixin_attributes |
+              comment          |
+              attribute        |
+              blank_attribute
+            ).repeat(0).as(:properties) >>
             str("}") >>
             space?
           ).as(:ruleset)
@@ -58,8 +76,22 @@ module Csscss
           ).as(:nested_ruleset)
         }
 
+        rule(:import) {
+          (
+            stri("@import") >>
+            match["^;"].repeat(1) >>
+            str(";") >>
+            space?
+          ).as(:import)
+        }
+
         rule(:blocks) {
-          space? >> (comment | nested_ruleset | ruleset).repeat(1).as(:blocks) >> space?
+          space? >> (
+            comment        |
+            import         |
+            nested_ruleset |
+            ruleset
+          ).repeat(1).as(:blocks) >> space?
         }
 
         root(:blocks)
@@ -69,6 +101,9 @@ module Csscss
         rule(nested_ruleset: sequence(:rulesets)) {
           rulesets
         }
+
+        rule(import: simple(:import)) { [] }
+        rule(mixin: simple(:mixin)) { nil }
 
         rule(comment: simple(:comment)) { nil }